% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_data.R
\docType{data}
\name{class_ipums_data}
\alias{class_ipums_data}
\title{An example dataset to be used with ipumsEDbook. Data comes from US 2015 ACS
1-year survey, harmonized by IPUMS.}
\format{
A data.frame of 3M records and 17 variables.
}
\source{
<https://international.ipums.org/international/index.shtml>
}
\usage{
data(clasS_ipums_data)
}
\description{
\describe{
  \item{COUNTRY}{Country Name}
  \item{YEAR}{Year of survey}
  \item{SAMPLE}{IPUMS numeric code of COUNTRY+YEAR}
  \item{SERIAL}{IPUMS unique hh identifier}
  \item{PERNUM}{Sequential number of individuals within each HH}
  \item{HHWT, PERWT}{Statistical weights}
  \item{RELATE, RELATED}{Relationship to head of house}
  \item{AGE, AGE2}{Age as integer or in 5-year bins}
  \item{SEX}{Reported biological sex}
  \item{MARST, MARSTD}{Marital status}
  \item{EDATTAIN, EDATTAIND}{Educational Attainment}
  \item{INCTOT}{Total income reported}
}
}
\keyword{datasets}
