% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_data.R
\docType{data}
\name{class_ipums_ddi}
\alias{class_ipums_ddi}
\title{Codebook to used with ipumsEDbook class_ipums_data. Data comes from US 2015 ACS
1-year survey, harmonized by IPUMS.}
\format{
## DDI for class_data
}
\source{
<https://international.ipums.org/international/index.shtml>
}
\usage{
data(clasS_ipums_ddi)
}
\description{
\describe{
  \item{file_name, file_path, file_type, file_encoding}{Technical info}
  \item{ipums_project, extract_date, extract_notes}{Technical info for extract system}
  \item{rectypes, rectypes_idvar, rectypes_keyvars}{Technical info of data structure}
  \item{var_info}{The actual Codebook}
  \item{conditions, citation}{Technical info for citations}
}
}
\keyword{datasets}
