% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_datasets.R
\docType{data}
\name{ex_census_data}
\alias{ex_census_data}
\alias{ex_census_data.rda}
\title{IPUMS census data}
\format{
## us2015
A data.frame of 3M records and 17 variables.
\describe{
  \item{COUNTRY}{Country Name}
  \item{YEAR}{Year of survey}
  \item{SAMPLE}{IPUMS numeric code of COUNTRY+YEAR}
  \item{SERIAL}{IPUMS unique hh identifier}
  \item{PERNUM}{Sequential number of individuals within each HH}
  \item{HHWT, PERWT}{Statistical weights}
  \item{RELATE, RELATED}{Relationship to head of house}
  \item{AGE, AGE2}{Age as integer or in 5-year bins}
  \item{SEX}{Reported biological sex}
  \item{MARST, MARSTD}{Marital status}
  \item{EDATTAIN, EDATTAIND}{Educational Attainment}
  \item{INCTOT}{Total income reported}
}
}
\source{
<https://international.ipums.org/international/index.shtml>
}
\description{
An example dataset to be used with ipumsEDbook. Data comes from US 2015 ACS
1-year survey, harmonized by IPUMS.
}
