% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_datasets.R
\docType{data}
\name{ex_height_weight}
\alias{ex_height_weight}
\title{Height and Weight data}
\format{
\describe{
  \item{AGE}{Country Name}
  \item{SEX}{Year of survey}
  \item{HEIGHT}{Height recorded in Inches (no shoes).}
  \item{WEIGHT}{Weight in lbs. Values 996,997,998,999 represent unknown data. Value 000 is NIU.}
}
}
\description{
An example dataset to be used with ipumsEDbook.
}
