% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_datasets.R
\docType{data}
\name{ex_wordcloud_wiki}
\alias{ex_wordcloud_wiki}
\title{Wordcloud wiki data}
\format{
\describe{
  \item{word}{List of each unique word.}
  \item{freq}{Frequency that word occurs in text.}
}
}
\description{
An example dataset to be used with ipumsEDbook. Frequencies of words from the
wikipedia page for 'data'.
}
